+++ orders/src/main/java/com/relatos_papel/orders/infrastructure/repositories/OrderItemRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface OrderItemRepository extends JpaRepository<OrderItem, Long> {
-    List<OrderItem> findByOrder_OrderId(Long orderId);
+public interface OrderItemRepository extends JpaRepository<OrderItem, Integer> {
+    List<OrderItem> findByOrder_OrderId(Integer orderId);
 }
